push_notif_agent/requirements.txt
PyYAML~=6.0
pytz~=2021.3
click~=8.0.3
Jinja2~=3.0.3
pip~=21.1.2
wheel~=0.36.2
cryptography~=36.0.1
six~=1.16.0
MarkupSafe~=2.0.1
certifi~=2021.10.8
urllib3~=1.26.7
idna~=3.3
ply~=3.11
pycparser~=2.21
pywebpush~=1.14.0
requests~=2.26.0
setuptools~=57.0.0